# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://www.baidu.com"
        },
        "version": "2.0"
    },
    "host": "192.168.1.163:7001",
    "paths": {
...
        "/v1/account/list": {
            "get": {
                "description": "获取股东数据列表分页 后台用户管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "获取股东数据列表分页",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id 可选参数",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取当页条数",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/account/pwd_login": {
            "post": {
                "description": "账号密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "账号密码登录",
                "parameters": [
                    {
                        "description": "账号密码登录 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.PwdLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/account/register": {
            "post": {
                "description": "股东账号注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "股东账号注册",
                "parameters": [
                    {
                        "description": "股东手机验证码登录 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.RegisterAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
        "/v1/shareholder/detail": {
            "get": {
                "description": "获取股东信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "获取股东信息",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/shareholder/income_list": {
            "get": {
                "description": "获取股东每日收益数据列表分页 后台用户管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "获取股东每日收益数据列表分页",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id 可选参数",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取当页条数",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/shareholder/info": {
            "get": {
                "description": "获取收益统计数据 app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "获取收益统计数据",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
        "/v1/shareholder/set_enable": {
            "post": {
                "description": "是否禁用用户 后台用户管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "是否禁用用户",
                "parameters": [
                    {
                        "description": "是否禁用用户 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.SetEnable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/shareholder/set_pwd": {
            "post": {
                "description": "股东设置提现密码 需要提供登录密码 后面可迭代加上短信验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "股东设置提现密码",
                "parameters": [
                    {
                        "description": "股东设置提现密码 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.SetWithdrawalPwd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/shareholder/subordinate_list": {
            "get": {
                "description": "获取股东下级列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "获取股东下级列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/shareholder/update_percent": {
            "post": {
                "description": "股东更新再分配占比例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东"
                ],
                "summary": "股东更新再分配占比例",
                "parameters": [
                    {
                        "description": "股东更新再分配占比例 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.UpdatePercentShareholderID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.PwdLogin": {
            "type": "object",
            "required": [
                "mobile",
                "pwd"
            ],
            "properties": {
                "mobile": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                }
            }
        },
        "request.RegisterAccount": {
            "type": "object",
            "required": [
                "is_enable",
                "mobile",
                "pwd",
                "sms_code"
            ],
            "properties": {
                "is_enable": {
                    "description": "是否开启",
                    "type": "boolean"
                },
                "mobile": {
                    "description": "手机号码",
                    "type": "string"
                },
                "pwd": {
                    "description": "密码",
                    "type": "string"
                },
                "recommend_code": {
                    "description": "推荐码 没有，可不填",
                    "type": "string"
                },
                "sms_code": {
                    "description": "短信验证码",
                    "type": "string"
                }
            }
        },
        "request.SetEnable": {
            "type": "object",
            "required": [
                "id",
                "is_enable",
                "shareholder_id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_enable": {
                    "type": "boolean"
                },
                "shareholder_id": {
                    "type": "string"
                }
            }
        },
        "request.SetWithdrawalPwd": {
            "type": "object",
            "required": [
                "login_pwd",
                "shareholder_id",
                "withdrawal_pwd"
            ],
            "properties": {
                "login_pwd": {
                    "type": "string"
                },
                "shareholder_id": {
                    "type": "string"
                },
                "withdrawal_pwd": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdatePercentShareholderID": {
            "type": "object",
            "required": [
                "shareholder_id"
            ],
            "properties": {
                "percent_five_shareholder_id": {
                    "type": "string"
                },
                "percent_three_shareholder_id": {
                    "type": "string"
                },
                "shareholder_id": {
                    "type": "string"
                }
            }
        },
...
        "response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}